requirements.txt
python-dotenv==0.20.0
requests==2.26.0
psycopg2-binary==2.9.3